Mastitis/requirements.txt
tensorflow
torch
torchvision
pillow
matplotlib
streamlit
numpy
